requirements.txt
cupy-cuda12x
imageio
matplotlib
networkx
numpy
pandas
pydantic
psycopg2
pyqtgraph
PySide6
pytest
pyyaml
ray
websockets
zstandard
